process manager refactoring
