+++ src/tracer/src/extracts/process/process_manager/mod.rs
@@ -1,10 +1,10 @@
 pub mod handlers;
 pub mod logger;
 pub mod matcher;
-pub mod process_manager;
+pub mod manager;
 pub mod metrics;
 pub mod state;
 pub mod system_refresher;
 
 // Re-export the main ProcessManager for convenience
-pub use process_manager::ProcessManager;
+pub use manager::ProcessManager;
